# services/academic-core-service/docs/openapi-front-office.yaml
...
openapi: 3.0.3
info:
  title: Academic Core Service – Front Office API
  description: |
    API for managing front office operations within the academic-core-service.
    This module handles day-to-day administrative activities including:
    - Visitors Log
    - Phone Call Log
    - Half Day Notices
    - Postal Dispatch/Receive
    - Admission Enquiries
    - Complaints
    
    All operations are scoped to a specific school and academic year.
  version: 1.0.0
servers:
  - url: /api/v1
    description: Base API path
...
paths:
...
  /front-office/phone-calls:
    get:
      summary: List phone calls
      description: |
        Retrieve a paginated list of phone call logs.
        Supports filtering by call type, date range, and follow-up date.
      tags: [Phone Calls]
      parameters:
        - $ref: '#/components/parameters/SchoolIdHeader'
        - $ref: '#/components/parameters/AcademicYearIdHeader'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
        - name: callType
          in: query
          schema:
            $ref: '#/components/schemas/CallTypeEnum'
          description: Filter by call type (INCOMING/OUTGOING)
        - name: fromDate
          in: query
          schema:
            type: string
            format: date
          description: Filter calls from this date
        - name: toDate
          in: query
          schema:
            type: string
            format: date
          description: Filter calls until this date
        - name: hasFollowUp
          in: query
          schema:
            type: boolean
          description: Filter calls with pending follow-up
        - name: search
          in: query
          schema:
            type: string
          description: Search by caller name or phone number
      responses:
        '200':
          description: Paginated list of phone calls
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhoneCallPageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Log phone call
      description: Record a new phone call entry.
      tags: [Phone Calls]
      parameters:
        - $ref: '#/components/parameters/SchoolIdHeader'
        - $ref: '#/components/parameters/AcademicYearIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePhoneCallRequest'
      responses:
        '201':
          description: Phone call logged successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhoneCallResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /front-office/phone-calls/{id}:
    get:
      summary: Get phone call details
      description: Retrieve details of a specific phone call log.
      tags: [Phone Calls]
      parameters:
        - $ref: '#/components/parameters/SchoolIdHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Phone call details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhoneCallResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # -------------------------------------------------------------------------
  # HALF DAY NOTICES
  # -------------------------------------------------------------------------
  /front-office/half-day-notices:
    get:
      summary: List half day notices
      description: |
        Retrieve a paginated list of half day notices.
        Supports filtering by class, section, and date.
      tags: [Half Day Notices]
      parameters:
        - $ref: '#/components/parameters/SchoolIdHeader'
        - $ref: '#/components/parameters/AcademicYearIdHeader'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
        - name: classId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by class
        - name: sectionId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by section
        - name: fromDate
          in: query
          schema:
            type: string
            format: date
          description: Filter notices from this date
        - name: toDate
          in: query
          schema:
            type: string
            format: date
          description: Filter notices until this date
        - name: studentId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by specific student
      responses:
        '200':
          description: Paginated list of half day notices
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HalfDayNoticePageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create half day notice
      description: Record a student's early departure.
      tags: [Half Day Notices]
      parameters:
        - $ref: '#/components/parameters/SchoolIdHeader'
        - $ref: '#/components/parameters/AcademicYearIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHalfDayNoticeRequest'
      responses:
        '201':
          description: Half day notice created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HalfDayNoticeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /front-office/half-day-notices/{id}:
    get:
      summary: Get half day notice details
      description: Retrieve details of a specific half day notice.
      tags: [Half Day Notices]
      parameters:
        - $ref: '#/components/parameters/SchoolIdHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Half day notice details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HalfDayNoticeResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # -------------------------------------------------------------------------
  # POSTAL DISPATCH / RECEIVE
  # -------------------------------------------------------------------------
  /front-office/postal-records:
    get:
      summary: List postal records
      description: |
        Retrieve a paginated list of postal dispatch and receive records.
        Supports filtering by direction, postal type, and date.
      tags: [Postal]
      parameters:
        - $ref: '#/components/parameters/SchoolIdHeader'
        - $ref: '#/components/parameters/AcademicYearIdHeader'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
        - name: direction
          in: query
          schema:
            $ref: '#/components/schemas/PostalDirectionEnum'
          description: Filter by direction (RECEIVED/DISPATCHED)
        - name: postalType
          in: query
          schema:
            $ref: '#/components/schemas/PostalTypeEnum'
          description: Filter by postal type (Letter/Parcel/Courier)
        - name: fromDate
          in: query
          schema:
            type: string
            format: date
          description: Filter records from this date
        - name: toDate
          in: query
          schema:
            type: string
            format: date
          description: Filter records until this date
        - name: search
          in: query
          schema:
            type: string
          description: Search by reference number, sender, or receiver
      responses:
        '200':
          description: Paginated list of postal records
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostalRecordPageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Add postal record
      description: Record a new postal dispatch or receive entry.
      tags: [Postal]
      parameters:
        - $ref: '#/components/parameters/SchoolIdHeader'
        - $ref: '#/components/parameters/AcademicYearIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostalRecordRequest'
      responses:
        '201':
          description: Postal record created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostalRecordResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /front-office/postal-records/{id}:
    get:
      summary: Get postal record details
      description: Retrieve details of a specific postal record.
      tags: [Postal]
      parameters:
        - $ref: '#/components/parameters/SchoolIdHeader'
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Postal record details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostalRecordResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # -------------------------------------------------------------------------
  # ADMISSION ENQUIRIES
  # -------------------------------------------------------------------------
...
components:
...
  parameters:
    SchoolIdHeader:
      name: X-School-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: The school context for the request
    AcademicYearIdHeader:
      name: X-Academic-Year-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: The academic year context for the request
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Page number (0-indexed)
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
    SortParam:
      name: sort
      in: query
      schema:
        type: string
      description: Sort criteria (e.g., 'createdAt,desc')

  # ---------------------------------------------------------------------------
  # RESPONSES
  # ---------------------------------------------------------------------------
  responses:
    BadRequest:
      description: Bad request - validation errors
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  # ---------------------------------------------------------------------------
  # SCHEMAS
  # ---------------------------------------------------------------------------
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
    PageMetadata:
      type: object
      properties:
        page:
          type: integer
          description: Current page number (0-indexed)
        size:
          type: integer
          description: Number of items per page
        totalElements:
          type: integer
          format: int64
          description: Total number of elements
        totalPages:
          type: integer
          description: Total number of pages

    # --- Enums ---
    FrontOfficeStatusEnum:
      type: string
      enum: [ACTIVE, CLOSED, ARCHIVED]
      description: Status for front office records
    CallTypeEnum:
      type: string
      enum: [INCOMING, OUTGOING]
      description: Type of phone call
    PostalDirectionEnum:
      type: string
      enum: [RECEIVED, DISPATCHED]
      description: Direction of postal record
    PostalTypeEnum:
      type: string
      enum: [LETTER, PARCEL, COURIER]
      description: Type of postal item
...
    PhoneCallResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        callerName:
          type: string
        phoneNumber:
          type: string
        callDate:
          type: string
          format: date
        callType:
          $ref: '#/components/schemas/CallTypeEnum'
        callDuration:
          type: integer
          description: Duration in minutes
          nullable: true
        description:
          type: string
          nullable: true
        nextFollowUpDate:
          type: string
          format: date
          nullable: true
        status:
          $ref: '#/components/schemas/FrontOfficeStatusEnum'
        remarks:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
    CreatePhoneCallRequest:
      type: object
      required: [callerName, phoneNumber, callDate, callType]
      properties:
        callerName:
          type: string
          minLength: 2
          maxLength: 100
        phoneNumber:
          type: string
          pattern: '^\+?[0-9]{10,15}$'
        callDate:
          type: string
          format: date
        callType:
          $ref: '#/components/schemas/CallTypeEnum'
        callDuration:
          type: integer
          minimum: 0
          description: Duration in minutes
        description:
          type: string
          maxLength: 1000
        nextFollowUpDate:
          type: string
          format: date
        remarks:
          type: string
          maxLength: 500
    PhoneCallPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/PhoneCallResponse'
        page:
          $ref: '#/components/schemas/PageMetadata'

    # --- Half Day Notices ---
    HalfDayNoticeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        studentId:
          type: string
          format: uuid
        studentName:
          type: string
          description: Denormalized for display
        classId:
          type: string
          format: uuid
        className:
          type: string
          description: Denormalized for display
        sectionId:
          type: string
          format: uuid
        sectionName:
          type: string
          description: Denormalized for display
        outTime:
          type: string
          format: date-time
        reason:
          type: string
        guardianName:
          type: string
        guardianPhone:
          type: string
        status:
          $ref: '#/components/schemas/FrontOfficeStatusEnum'
        remarks:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
    CreateHalfDayNoticeRequest:
      type: object
      required: [studentId, classId, sectionId, outTime, reason, guardianName, guardianPhone]
      properties:
        studentId:
          type: string
          format: uuid
        classId:
          type: string
          format: uuid
        sectionId:
          type: string
          format: uuid
        outTime:
          type: string
          format: date-time
        reason:
          type: string
          minLength: 2
          maxLength: 500
        guardianName:
          type: string
          minLength: 2
          maxLength: 100
        guardianPhone:
          type: string
          pattern: '^\+?[0-9]{10,15}$'
        remarks:
          type: string
          maxLength: 500
    HalfDayNoticePageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/HalfDayNoticeResponse'
        page:
          $ref: '#/components/schemas/PageMetadata'

    # --- Postal Records ---
    PostalRecordResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        direction:
          $ref: '#/components/schemas/PostalDirectionEnum'
        postalType:
          $ref: '#/components/schemas/PostalTypeEnum'
        referenceNumber:
          type: string
          nullable: true
        fromTitle:
          type: string
        toTitle:
          type: string
        courierName:
          type: string
          nullable: true
        date:
          type: string
          format: date
        attachmentUrl:
          type: string
          format: uri
          nullable: true
        notes:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/FrontOfficeStatusEnum'
        remarks:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
    CreatePostalRecordRequest:
      type: object
      required: [direction, postalType, fromTitle, toTitle, date]
      properties:
        direction:
          $ref: '#/components/schemas/PostalDirectionEnum'
        postalType:
          $ref: '#/components/schemas/PostalTypeEnum'
        referenceNumber:
          type: string
          maxLength: 100
        fromTitle:
          type: string
          minLength: 2
          maxLength: 200
        toTitle:
          type: string
          minLength: 2
          maxLength: 200
        courierName:
          type: string
          maxLength: 100
        date:
          type: string
          format: date
        attachmentUrl:
          type: string
          format: uri
        notes:
          type: string
          maxLength: 1000
        remarks:
          type: string
          maxLength: 500
    PostalRecordPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/PostalRecordResponse'
        page:
          $ref: '#/components/schemas/PageMetadata'

    # --- Admission Enquiries ---
...